Fix all phpstan level 4 and a few level 5 issues

+++ src/Core/Checkout/Cart/Delivery/DeliveryBuilder.php
@@ -82,7 +82,7 @@ class DeliveryBuilder
                 continue;
             }
 
-            $outOfStock = (int) abs($item->getDeliveryInformation()->getStock() - $quantity);
+            $outOfStock = abs($item->getDeliveryInformation()->getStock() - $quantity);
 
             $position = $this->recalculatePosition(
                 $item,
